+++ CHANGES.md
@@ -1,5 +1,13 @@
 ### (unreleased)
 
+#### Changes
+
+  + Add the option `doc-comments-val=before|after` (#1012) (Jules Aguillon)
+    This option set the placement of documentation comment on `val` and `external` only.
+    It is set to `after` by default.
+    The default for `doc-comments` is changed from `after` to `before`, this
+    option will be deleted in the future.
+
 #### Bug fixes
 
   + Fix missing parentheses around `let open` (#1229) (Jules Aguillon)